requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
deprecation==2.1.0
fastapi==0.104.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
joblib==1.5.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
postgrest==2.21.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
realtime==2.21.1
requests==2.31.0
rich==14.1.0
scikit-learn==1.2.2
scipy==1.16.2
seaborn==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
# tensorflow_intel==2.18.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
